Move view amount format  into ViewAmountFormat

+++ app/src/main/java/ua/com/radiokot/money/accounts/data/PowerSyncAccountRepository.kt
@@ -71,7 +71,7 @@ class PowerSyncAccountRepository(
             id = getString(column)!!,
             code = getString(++column)!!.trim(),
             symbol = getString(++column)!!.trim(),
-            precision = getLong(++column)!!.toShort(),
+            precision = getLong(++column)!!.toInt(),
         )
 
         Account(
